Run with validateDrp.py.  Refactor and rename code files.
Simplified names of (check,plot)AstrometryPhotometry.py to check.py, plot.py
Added print.py to handle printing.
Include YAML example file for CFHT.

Can now run with, e.g.:
   validateDrp.py CFHT/output examples/runCfht.yaml
