+++ examples/runCfht.yaml
