+++ examples/runDecam.yaml
@@ -4,7 +4,7 @@
 good_mag_limit: 21.0  # Only consider stars bright than this limit
 medianAstromscatterRef: 25  # expected astrometric variability [mas]
 medianPhotoscatterRef: 25  # expected photometric variability [mmag]
-matchRef: 5000  # Number of stars expected to match.  Update if you change the ccd list below
-visits: [849375, 850587]
-filter: 'r'
-ccd: [12, 13, 14, 21, 22, 23]
+matchRef: 10000  # Number of stars expected to match.  Update if you change the ccd list below
+visits: [176837, 176846]
+filter: 'z'
+ccd: [10, 11, 12, 13, 14, 15, 16, 17, 18]
