+++ python/lsst/validate/drp/plot.py
